storage/fs/{,cocoonfs}: widen inode numbers to 64 bits
From discussions ([1]) in the context of the "COCONUT Secure VM Service
Module" ([2]), it has become clear that an inode number space of 32
bits is a bit scarce for partitioning it statically into
per-application/-service ranges. For example, a vTPM implementation
alone might very well be making use of four full 24 bit inode number
subspaces with the most significant byte set to either 0x01, 0x11, 0x12
(NV indices) or 0x81 (persistent objects) already, and that list can
potentially grow with future TPM spec updates.

As we don't have a crystal ball, widen the CocoonFs inode number space
to 64 bits. With that, a dedicated 32 bit subspace can get allocated
statically to each SVSM application/service, of which there are plenty
(2^32), and which should be large enough to satisfy each individual
application's needs.

[1] https://github.com/coconut-svsm/svsm/pull/806
[2] https://github.com/coconut-svsm/svsm

+++ cocoonfs-cli/src/main.rs
@@ -229,8 +229,8 @@ struct CliWriteFileArgs {
     enable_trimming: bool,
 
     /// Inode number of the file to write to.
-    #[arg(value_name="INODE-NUMBER", value_parser = clap::value_parser!(u32).range(6..))]
-    inode: u32,
+    #[arg(value_name="INODE-NUMBER", value_parser = clap::value_parser!(u64).range(6..))]
+    inode: u64,
 }
 
 #[derive(clap::Args)]
@@ -247,8 +247,8 @@ struct CliReadFileArgs {
     enable_trimming: bool,
 
     /// Inode number of the file to read from.
-    #[arg(value_name="INODE-NUMBER", value_parser = clap::value_parser!(u32).range(6..))]
-    inode: u32,
+    #[arg(value_name="INODE-NUMBER", value_parser = clap::value_parser!(u64).range(6..))]
+    inode: u64,
 }
 
 #[derive(clap::Args)]
@@ -271,8 +271,8 @@ struct CliRemoveFile {
     enable_trimming: bool,
 
     /// Inode number to delete.
-    #[arg(value_name="INODE-NUMBER", value_parser = clap::value_parser!(u32).range(6..))]
-    inode: u32,
+    #[arg(value_name="INODE-NUMBER", value_parser = clap::value_parser!(u64).range(6..))]
+    inode: u64,
 }
 
 #[derive(clap::Args)]
@@ -1047,7 +1047,7 @@ fn main() {
                 NvFsReadContext::Committed {
                     seq: consistent_read_sequence,
                 },
-                6..=u32::MAX,
+                6..=u64::MAX,
             ) {
                 Ok(Ok(enumerate_cursor)) => enumerate_cursor,
                 Ok(Err((_, e))) | Err(e) => {
