+++ storage/src/fs/cocoonfs/test/mod.rs
@@ -395,7 +395,7 @@ fn cocoonfs_test_commit_transaction_op_helper(
 fn cocoonfs_test_write_inode_op_helper(
     fs_instance: &<TestCocoonFs as fs::NvFs>::SyncRcPtr,
     transaction: <TestCocoonFs as fs::NvFs>::Transaction,
-    inode: u32,
+    inode: u64,
     data: &[u8],
 ) -> Result<<TestCocoonFs as fs::NvFs>::Transaction, fs::NvFsError> {
     let rng = Box::new(rng::test_rng());
@@ -419,7 +419,7 @@ fn cocoonfs_test_write_inode_op_helper(
 fn cocoonfs_test_read_inode_op_helper(
     fs_instance: &<TestCocoonFs as fs::NvFs>::SyncRcPtr,
     read_context: Option<fs::NvFsReadContext<TestCocoonFs>>,
-    inode: u32,
+    inode: u64,
 ) -> Result<(fs::NvFsReadContext<TestCocoonFs>, Option<zeroize::Zeroizing<Vec<u8>>>), fs::NvFsError> {
     let rng = Box::new(rng::test_rng());
     let read_inode_fut =
@@ -437,14 +437,14 @@ fn cocoonfs_test_read_inode_op_helper(
 fn cocoonfs_test_enumerate_inodes_op_collect(
     fs_instance: &<TestCocoonFs as fs::NvFs>::SyncRcPtr,
     read_context: Option<fs::NvFsReadContext<TestCocoonFs>>,
-    inodes_enumerate_range: ops::RangeInclusive<u32>,
-) -> Result<(fs::NvFsReadContext<TestCocoonFs>, Vec<u32>), fs::NvFsError> {
+    inodes_enumerate_range: ops::RangeInclusive<u64>,
+) -> Result<(fs::NvFsReadContext<TestCocoonFs>, Vec<u64>), fs::NvFsError> {
     struct CollectInodesCallback {
-        collected_inodes: Vec<u32>,
+        collected_inodes: Vec<u64>,
     }
 
     impl CocoonFsTestEnumerateInodesFutureCallback for CollectInodesCallback {
-        fn call(&mut self, inode: u32, _inode_data: zeroize::Zeroizing<Vec<u8>>) -> Result<(), fs::NvFsError> {
+        fn call(&mut self, inode: u64, _inode_data: zeroize::Zeroizing<Vec<u8>>) -> Result<(), fs::NvFsError> {
             self.collected_inodes.push(inode);
             Ok(())
         }
@@ -465,7 +465,7 @@ fn cocoonfs_test_enumerate_inodes_op_collect(
 fn cocoonfs_test_enumerate_inodes_op_cb<CB: CocoonFsTestEnumerateInodesFutureCallback>(
     fs_instance: &<TestCocoonFs as fs::NvFs>::SyncRcPtr,
     read_context: Option<fs::NvFsReadContext<TestCocoonFs>>,
-    inodes_enumerate_range: ops::RangeInclusive<u32>,
+    inodes_enumerate_range: ops::RangeInclusive<u64>,
     callback: CB,
 ) -> Result<(fs::NvFsReadContext<TestCocoonFs>, CB), fs::NvFsError> {
     let rng = Box::new(rng::test_rng());
@@ -486,7 +486,7 @@ fn cocoonfs_test_enumerate_inodes_op_cb<CB: CocoonFsTestEnumerateInodesFutureCal
 }
 
 trait CocoonFsTestEnumerateInodesFutureCallback: 'static + marker::Unpin + marker::Send {
-    fn call(&mut self, inode: u32, inode_data: zeroize::Zeroizing<Vec<u8>>) -> Result<(), fs::NvFsError>;
+    fn call(&mut self, inode: u64, inode_data: zeroize::Zeroizing<Vec<u8>>) -> Result<(), fs::NvFsError>;
 }
 
 struct CocoonFsTestEnumerateInodesFuture<CB: CocoonFsTestEnumerateInodesFutureCallback> {
@@ -499,19 +499,19 @@ struct CocoonFsTestEnumerateInodesFuture<CB: CocoonFsTestEnumerateInodesFutureCa
 enum CocoonFsTestEnumerateInodesFutureState {
     StartReadSequence {
         start_read_sequence_fut: <TestCocoonFs as fs::NvFs>::StartReadSequenceFut,
-        inodes_enumerate_range: ops::RangeInclusive<u32>,
+        inodes_enumerate_range: ops::RangeInclusive<u64>,
     },
     CreateEnumerateCursor {
         // Is mandatory, lives in an Option<> only so that it can be taken out of a mutable reference on
         // Self.
         read_context: Option<fs::NvFsReadContext<TestCocoonFs>>,
-        inodes_enumerate_range: ops::RangeInclusive<u32>,
+        inodes_enumerate_range: ops::RangeInclusive<u64>,
     },
     Next {
         next_fut: <<TestCocoonFs as fs::NvFs>::EnumerateCursor as fs::NvFsEnumerateCursor<TestCocoonFs>>::NextFut,
     },
     ReadCurrentInodeData {
-        inode: u32,
+        inode: u64,
         read_fut:
             <<TestCocoonFs as fs::NvFs>::EnumerateCursor as fs::NvFsEnumerateCursor<TestCocoonFs>>::ReadInodeDataFut,
     },
@@ -522,7 +522,7 @@ impl<CB: CocoonFsTestEnumerateInodesFutureCallback> CocoonFsTestEnumerateInodesF
     fn new(
         fs_instance: &<TestCocoonFs as fs::NvFs>::SyncRcPtr,
         read_context: Option<fs::NvFsReadContext<TestCocoonFs>>,
-        inodes_enumerate_range: ops::RangeInclusive<u32>,
+        inodes_enumerate_range: ops::RangeInclusive<u64>,
         callback: CB,
     ) -> Result<Self, fs::NvFsError> {
         let fut_state = match read_context {
@@ -693,7 +693,7 @@ impl<CB: CocoonFsTestEnumerateInodesFutureCallback> fs::NvFsFuture<TestCocoonFs>
 fn cocoonfs_test_unlink_inodes_op_uncond(
     fs_instance: &<TestCocoonFs as fs::NvFs>::SyncRcPtr,
     transaction: <TestCocoonFs as fs::NvFs>::Transaction,
-    inodes_unlink_range: ops::RangeInclusive<u32>,
+    inodes_unlink_range: ops::RangeInclusive<u64>,
 ) -> Result<<TestCocoonFs as fs::NvFs>::Transaction, fs::NvFsError> {
     cocoonfs_test_unlink_inodes_op_fnmut_cb(fs_instance, transaction, inodes_unlink_range, |_inode, _inode_data| {
         Ok(true)
@@ -702,23 +702,23 @@ fn cocoonfs_test_unlink_inodes_op_uncond(
 }
 
 fn cocoonfs_test_unlink_inodes_op_fnmut_cb<
-    CB: 'static + FnMut(u32, &[u8]) -> Result<bool, fs::NvFsError> + marker::Send + marker::Unpin,
+    CB: 'static + FnMut(u64, &[u8]) -> Result<bool, fs::NvFsError> + marker::Send + marker::Unpin,
 >(
     fs_instance: &<TestCocoonFs as fs::NvFs>::SyncRcPtr,
     transaction: <TestCocoonFs as fs::NvFs>::Transaction,
-    inodes_unlink_range: ops::RangeInclusive<u32>,
+    inodes_unlink_range: ops::RangeInclusive<u64>,
     callback: CB,
 ) -> Result<(<TestCocoonFs as fs::NvFs>::Transaction, CB), fs::NvFsError> {
     struct UnlinkInodesCallback<
-        CB: 'static + FnMut(u32, &[u8]) -> Result<bool, fs::NvFsError> + marker::Send + marker::Unpin,
+        CB: 'static + FnMut(u64, &[u8]) -> Result<bool, fs::NvFsError> + marker::Send + marker::Unpin,
     > {
         callback: CB,
     }
 
-    impl<CB: 'static + FnMut(u32, &[u8]) -> Result<bool, fs::NvFsError> + marker::Send + marker::Unpin>
+    impl<CB: 'static + FnMut(u64, &[u8]) -> Result<bool, fs::NvFsError> + marker::Send + marker::Unpin>
         CocoonFsTestUnlinkInodesFutureCallback for UnlinkInodesCallback<CB>
     {
-        fn call(&mut self, inode: u32, inode_data: zeroize::Zeroizing<Vec<u8>>) -> Result<bool, fs::NvFsError> {
+        fn call(&mut self, inode: u64, inode_data: zeroize::Zeroizing<Vec<u8>>) -> Result<bool, fs::NvFsError> {
             (self.callback)(inode, inode_data.as_slice())
         }
     }
@@ -736,7 +736,7 @@ fn cocoonfs_test_unlink_inodes_op_fnmut_cb<
 fn cocoonfs_test_unlink_inodes_op_cb<CB: CocoonFsTestUnlinkInodesFutureCallback>(
     fs_instance: &<TestCocoonFs as fs::NvFs>::SyncRcPtr,
     transaction: <TestCocoonFs as fs::NvFs>::Transaction,
-    inodes_unlink_range: ops::RangeInclusive<u32>,
+    inodes_unlink_range: ops::RangeInclusive<u64>,
     callback: CB,
 ) -> Result<(<TestCocoonFs as fs::NvFs>::Transaction, CB), fs::NvFsError> {
     let rng = Box::new(rng::test_rng());
@@ -757,7 +757,7 @@ fn cocoonfs_test_unlink_inodes_op_cb<CB: CocoonFsTestUnlinkInodesFutureCallback>
 }
 
 trait CocoonFsTestUnlinkInodesFutureCallback: 'static + marker::Unpin + marker::Send {
-    fn call(&mut self, inode: u32, inode_data: zeroize::Zeroizing<Vec<u8>>) -> Result<bool, fs::NvFsError>;
+    fn call(&mut self, inode: u64, inode_data: zeroize::Zeroizing<Vec<u8>>) -> Result<bool, fs::NvFsError>;
 }
 
 struct CocoonFsTestUnlinkInodesFuture<CB: CocoonFsTestUnlinkInodesFutureCallback> {
@@ -772,7 +772,7 @@ enum CocoonFsTestUnlinkInodesFutureState {
         next_fut: <<TestCocoonFs as fs::NvFs>::UnlinkCursor as fs::NvFsUnlinkCursor<TestCocoonFs>>::NextFut,
     },
     ReadCurrentInodeData {
-        inode: u32,
+        inode: u64,
         read_fut: <<TestCocoonFs as fs::NvFs>::UnlinkCursor as fs::NvFsUnlinkCursor<TestCocoonFs>>::ReadInodeDataFut,
     },
     UnlinkCurrentInode {
@@ -785,7 +785,7 @@ impl<CB: CocoonFsTestUnlinkInodesFutureCallback> CocoonFsTestUnlinkInodesFuture<
     fn new(
         fs_instance: &<TestCocoonFs as fs::NvFs>::SyncRcPtr,
         transaction: <TestCocoonFs as fs::NvFs>::Transaction,
-        inodes_unlink_range: ops::RangeInclusive<u32>,
+        inodes_unlink_range: ops::RangeInclusive<u64>,
         callback: CB,
     ) -> Result<Self, fs::NvFsError> {
         let unlink_cursor = <TestCocoonFs as fs::NvFs>::unlink_cursor(
